Adding sync version of CfuWriter (#21)
This PR adds a sync version of `CfuWriter` and merged `CfuWriterMut`
with `CfuWriter`.

- The previous `CfuWriter` is now named `CfuWriterAsync` and the
synchronous version is named `CfuWriterSync`. To keep the previous
behavior, use `CfuWriterAsync`.
- Rename the Default Writer to `CfuWriterNop` to be more explicit about
what it does.
- Removed default implementations that were not default behavior. The
reason for the change, add clarity to the user about which function
needs to be implemented to have a working writer and will "force" the
use to implement these functions. In the event where someone wants a
dummy implementation, the intent will be clearer now that it is moved
where the implementation is.

# Issues

Fix #5  
Fix #17